Start work on a gdigi API.
The API allows a client to exercise the gui via simple get/set commands. It's
implemented via a unix domain socket and hooked into the message loop.

The 'get' command allows a client (or clients) to send a
'REQUEST_PARAMETER_VALUE' to the device. When the device responds, the result
is returned to the client.

The 'set' command sends a 'RECEIVE_PARAMETER_VALUE' to set a parameter to the
value specified by the client. This is then followed by a
'REQUEST_PARAMETER_VALUE' so that the gui is exercised.

Code changes:

1) New header files gdigi_api_private.h and gdigi_api.h for the server and
client sides of the API.

2) Implement the server side of the API in gdigi_api_server.c.

3) Add the new module to the makefile and fix a minor bug in distclean.

4) Hook the API server functions into gdigi.c.

5) In the message thread, reduce the poll timeout to 50 from 200 ms to improve
   API RTT.

6) Add a new debug flag for the API.

7 )Add a macro for the commonly used operation of packing an id and position
   into an integer and then casting it to a gpointer.

8) Add the Gdigi.pm perl module. This implements the client side of the API in
   perl.  Eventually, there will be a C library used to build a C client and a
   proper perl extension.

9) Add a simple perl program gdigi.pl to exercise the API via the perl module.